# Optimize `is_bogon` by 10-100x