@Alextopher 2024-01-02
This change should be semver patch compatible.